requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
discord.py==2.4.0
httpx==0.27.2
